refactor(css): fold the value-syntax parser into the data generator
The parser had one consumer, and its only other reader was its own test. It
now lives beside the collectors that walk its trees, and the generation runs
only when the file is the entry point — so requiring it for the parser exports
parses no grammar.
