+++ lib/css/syntax.js
@@ -6302,7 +6302,7 @@ const _NESTED_TERM_RE = /^(?:\d*\.\d+|\d+)(?:%|[a-z]+)?$/i;
  * `width:0` is not.
  * TODO take the parentheses off a negative too where the property accepts one
  * (`margin-left:calc(-5px)` is `-5px`). The grammars annotate their ranges
- * (`<length [0,∞]>`), which `tooling/css-value-syntax.js` already parses into
+ * (`<length [0,∞]>`), which `tooling/generate-css-data.js` already parses into
  * `TypeNode.min`, so the set is derivable the way `INTEGER_PROPERTIES` is.
  * @param {string} fragment one printed component
  * @param {string} property the lowercased property it belongs to
